Building out Manager Layout
